requirements.txt
alembic==1.18.4
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
cachelib==0.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
distro==1.9.0
Flask==3.0.2
Flask-Caching==2.3.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.3
openai==2.30.0
packaging==26.0
pandas==3.0.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
stripe==15.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.0.1
APScheduler